requirements.txt
Flask==1.1.2
IPython
plotly
azure.cosmos
dash
pandas
plotly.express
dash_bootstrap_components
azure.ai.textanalytics
azure.core
pyyaml
itsdangerous==2.0.1
